Add basic privacy policy
App does almost nothing that deals with personal data of any form, but
it's good to have a privacy policy that explicitly states this.

Previously the play store entry pointed to the privacy policy for
PostgreSQL Europe which would only be valid when used against
conferences actually run by PGEU. In preparation for public launch, it's
better to repoint it to this more generic one.