Fix random-rows-are-bold in statistics view
Due to how android recycles the views in a listview (or in this case an
expandable listview), random rows in the subsections could end up being
boldface, instead of just the summary rows as intended.

Create two different types of chlid views in the list, one for regular
rows and one for bold rows, to get around this. They can still use the
same layout.