Wrap all layouts in a DrawerLayout, maybe fixes padding?
Padding fixed only part of the problem, and the status bar was still
broken. Instead, wrap everything in a DrawerLayout (where for some
reason the drawer doesn't appear, but let's ignore that) which appears
to do the right thing with the same parameters.

Hopefully this also fixes the buttons going off-screen when using soft
buttons.